update code add WorkerVO

+++ back-work/src/main/java/kr/co/dbinc/back_work/model/WorkVO.java
@@ -16,13 +16,9 @@ public class WorkVO {
 	private String workTitle;
 	private String workContent;
 	private Integer workState;
-	private Integer workerID;
+	private Long workerID;
 
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private Date startDate;
-	private Date finishDate;
-
-	
-
-	
+	private Date finishDate;	
 }
